libcamera is moving its IRC channel to OFTC (irc.oftc.net)
As reported by lwn.net in https://lwn.net/Articles/856543/ and
https://lwn.net/Articles/857140/, freenode, the largest IRC network and
the home of our IRC channel, is embroiled in a messy dispute. After
careful consideration, we have decided that it is not a suitable place
to host our community anymore.

Effective immediately, the #libcamera IRC channel is transferred to the
OFTC IRC network. We will continue our effort to make the channel a
welcoming place for everybody.

Usage of OFTC is governed by its network [1] and privacy [2] policies.
While we don't expect these to cause issues for anyone, please feel free
to contact OFTC or libcamera should you have any question or concern.

[1] https://www.oftc.net/Network_Policy/
[2] https://www.oftc.net/Privacy_Policy/

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Umang Jain <umang.jain@ideasonboard.com>