pipeline: ipu3: Report error when failing to configure
If the IPA fails to configure, this can now be caught by the pipeline handler
but the cause may not be clear.

If the IPA is isolated, then reports from that IPA will not be visible
in the libcamera logs directly.

Print the return error value to help identify any issue that has arisen.

Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Umang Jain <umang.jain@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>