android: cros: Simplify integration
The ChromeOS specific Android HAL support was overly-complex to support
linking an additional library, and then passing the built object into
the libcamera.so build.

Now that it has its own distinct libcamera-hal.so, simplify the CrOS
integration to build directly into that library.

The removal of the -Wno-shadow here is intentional, as it is not
required for compilation of the camera3_hal.o object.

Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>