android: capabilities: Collect per-stream frame durations
Collect the per-stream frame durations while building the list
of supported stream formats and resolutions.

In order to get an updated list of controls it is necessary to apply
to the Camera the configuration we're testing, which was so far only
validated.

The per-configuration durations will be used to populate the Android
ANDROID_SCALER_AVAILABLE_MIN_FRAME_DURATIONS static metadata.

Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Umang Jain <umang.jain@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>